+++ docs/include/samplecode/SoundTouchClient/SetAudioProductToneControls.py
@@ -13,9 +13,9 @@ try:
     # note that not all devices support retrieval of this information.
     cfgBefore:AudioProductToneControls = None
     cfgBefore = client.GetAudioProductToneControls()
-    print("\nCurrent Audio Product Tone Control Levels: \n%s" % cfgBefore.ToString())
+    print("\nCurrent audio product tone controls: \n%s" % cfgBefore.ToString())
         
-    # create new tone controls object.
+    # create new audio product tone controls object.
     cfgUpdate:AudioProductToneControls = AudioProductToneControls()
 
     # for testing purposes, toggle the Bass level.  
@@ -23,21 +23,21 @@ try:
     cfgUpdate.Bass.Value = cfgBefore.Bass.MinValue
     if cfgUpdate.Bass.Value == cfgBefore.Bass.Value:
         cfgUpdate.Bass.Value = cfgBefore.Bass.MaxValue
-    print("\nSetting Audio Product Tone Control Bass Level to '%s' (from '%s') ..." % (cfgUpdate.Bass.Value, cfgBefore.Bass.Value))
+    print("\nSetting audio product tone controls Bass Level to '%s' (from '%s') ..." % (cfgUpdate.Bass.Value, cfgBefore.Bass.Value))
                 
     # for testing purposes, toggle the Treble level.  
     # if the level is currently minValue, then we will set to maxValue.
     cfgUpdate.Treble.Value = cfgBefore.Treble.MinValue
     if cfgUpdate.Treble.Value == cfgBefore.Treble.Value:
         cfgUpdate.Treble.Value = cfgBefore.Treble.MaxValue
-    print("Setting Audio Product Tone Control Treble Level to '%s' (from '%s') ..." % (cfgUpdate.Treble.Value, cfgBefore.Treble.Value))
+    print("Setting audio product tone controls Treble Level to '%s' (from '%s') ..." % (cfgUpdate.Treble.Value, cfgBefore.Treble.Value))
                 
     # update audio product tone controls.
     client.SetAudioProductToneControls(cfgUpdate)
             
     # get current audio product tone controls.
     cfgAfter:AudioProductToneControls = client.GetAudioProductToneControls(True)
-    print("\nChanged Audio Product Tone Controls: \n%s" % (cfgAfter.ToString()))
+    print("\nChanged audio product tone controls: \n%s" % (cfgAfter.ToString()))
         
 except Exception as ex:
 
@@ -46,10 +46,10 @@ except Exception as ex:
 finally:
     
     if cfgBefore is not None:
-        # reset audio product tone controls to original values.
-        print("\nRestoring Audio Product Tone Controls to original values.")
+        
+        # restore audio product tone controls to original values.
         client.SetAudioProductToneControls(cfgBefore)            
 
         # get current audio product tone controls.
         cfgAfter:AudioProductToneControls = client.GetAudioProductToneControls(True)
-        print("\nRestored Audio Product Tone Controls: \n%s" % (cfgAfter.ToString()))
+        print("\nRestored audio product tone controls: \n%s" % (cfgAfter.ToString()))